More complete implementation of loadshader: read all the ops & args.
git-svn-id: https://openshadinglanguage.googlecode.com/svn/trunk@70 a3683778-fedd-11de-87c3-653bb6514e15

+++ src/liboslcomp/oslcomp.cpp
@@ -443,7 +443,7 @@ OSLCompilerImpl::write_oso_file (const std::string &outfilename)
         }
 
         // Jump targets
-        for (int i = 0;  i < Opcode::max_jumps;  ++i)
+        for (size_t i = 0;  i < Opcode::max_jumps;  ++i)
             if (op->jump(i) >= 0)
                 oso ("%d ", op->jump(i));
